+++ packages/unmock-core/src/index.ts
@@ -1,10 +1,12 @@
 // Sinon for asserts and matchers
 import * as sinon from "sinon";
 import Backend, { buildRequestHandler } from "./backend";
+import UnmockFaker from "./faker";
 import { ILogger, IUnmockOptions, IUnmockPackage } from "./interfaces";
-import { ExtendedJSONSchema, nockify, vanillaJSONSchemify } from "./nock";
 import internalRunner, { IRunnerOptions } from "./runner";
+import { addFromNock, NockAPI, ServiceStore } from "./service/serviceStore";
 import { AllowedHosts, BooleanSetting, IBooleanSetting } from "./settings";
+import * as typeUtils from "./utils";
 
 export * from "./interfaces";
 export * from "./types";
@@ -14,6 +16,8 @@ export { transform, Addl, Arr } from "./generator-utils";
 export { IService } from "./service/interfaces";
 export { ServiceCore } from "./service/serviceCore";
 export { Backend, buildRequestHandler };
+export { typeUtils };
+export { UnmockFaker };
 
 export class UnmockPackage implements IUnmockPackage {
   public allowedHosts: AllowedHosts;
@@ -23,6 +27,8 @@ export class UnmockPackage implements IUnmockPackage {
    */
   public randomize: IBooleanSetting;
   public readonly backend: Backend;
+  public readonly nock: NockAPI;
+  private readonly opts: IUnmockOptions;
   private logger: ILogger = { log: () => undefined }; // Default logger does nothing
   constructor(
     backend: Backend,
@@ -36,16 +42,41 @@ export class UnmockPackage implements IUnmockPackage {
     this.allowedHosts = new AllowedHosts();
     this.useInProduction = new BooleanSetting(false);
     this.randomize = new BooleanSetting(false);
-  }
-
-  public on() {
-    const opts: IUnmockOptions = {
+    this.opts = {
       useInProduction: () => this.useInProduction.get(),
       isWhitelisted: (url: string) => this.allowedHosts.isWhitelisted(url),
       log: (message: string) => this.logger.log(message),
       randomize: () => this.randomize.get(),
     };
-    this.backend.initialize(opts);
+    this.nock = addFromNock(this.backend.serviceStore);
+  }
+
+  /**
+   * Create a new `UnmockFaker` with empty `ServiceStore`.
+   *
+   * @example
+   *
+   * const faker = unmock.faker();
+   * faker
+   *  .nock('https://api.github.com', 'github')
+   *  .get('/v1/users')
+   *  .reply({ id: '1' });
+   * const req: ISerializedRequest = {
+   *  host: "api.github.com",
+   *  path: '/v1/users',
+   *  protocol: 'https',
+   *  ...
+   * };
+   * const res = faker.generate(request);
+   *
+   * @returns UnmockFaker instance
+   */
+  public faker(): UnmockFaker {
+    return new UnmockFaker({ serviceStore: new ServiceStore([]) });
+  }
+
+  public on() {
+    this.backend.initialize(this.opts);
     return this;
   }
   public init() {
@@ -70,34 +101,6 @@ export class UnmockPackage implements IUnmockPackage {
     return f;
   }
 
-  public nock(
-    baseUrl: string,
-    nameOrHeaders?:
-      | string
-      | { reqheaders?: Record<string, ExtendedJSONSchema> },
-    name?: string,
-  ) {
-    const internalName =
-      typeof nameOrHeaders === "string"
-        ? nameOrHeaders
-        : typeof name === "string"
-        ? name
-        : undefined;
-    const requestHeaders =
-      typeof nameOrHeaders === "object" && nameOrHeaders.reqheaders
-        ? Object.entries(nameOrHeaders.reqheaders).reduce(
-            (a, b) => ({ ...a, [b[0]]: vanillaJSONSchemify(b[1]) }),
-            {},
-          )
-        : {};
-    return nockify({
-      backend: this.backend,
-      baseUrl,
-      requestHeaders,
-      name: internalName,
-    });
-  }
-
   public associate(url: string, name: string) {
     this.backend.serviceStore.updateOrAdd({ baseUrl: url, name });
   }
@@ -107,6 +110,6 @@ export class UnmockPackage implements IUnmockPackage {
   }
 
   public reset() {
-    Object.values(this.backend.services).forEach(service => service.reset());
+    this.backend.serviceStore.resetServices();
   }
 }