+++ packages/unmock-server/index.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-const main = require("./dist/index.js");
-main.run();