+++ packages/unmock/src/browser/index.ts
@@ -1,0 +1,4 @@
+export * from "unmock-core";
+import unmock from "unmock-browser";
+export * from "unmock-browser";
+export default unmock;